Inter Game/genai-intern-game/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
aioredis==2.0.1
pytest==7.4.3
httpx==0.25.2
alembic==1.12.1
python-magic==0.4.27 